Handle displaying all errors in cli main instead of in command run funcs
This consolidates (almost) all of the error display handling into the cli
main(), which is idiomatic since the library side of this shouldn't need to
worry about how its errors are presented.

There are a couple spots where the Error case is a Vec of things, which can't
be handled by the single error handler in main. These have been changed to
do their custom error printing, then return a new error to main(). I added
TODO comments to these.

Signed-off-by: Robert Detjens <github@detjens.dev>